backend/requirements.txt
Flask==0.12.4
pyjwt==1.7.1
flask-cors==3.0.7
google-auth==1.6.2
requests==2.21.0
requests-toolbelt==0.9.1
google-cloud-firestore==0.32.1

